ignore more tests and renamed test file to use kebab-case
